+++ src/scene/TextureScene.h
@@ -20,7 +20,7 @@ namespace GL {
 
         ~TextureScene() override = default;
 
-        void OnUpdate(float deltaTime, GLFWwindow *window) override;
+        void OnUpdate(double deltaTime, GLFWwindow *window) override;
 
         void OnRender() override;
